Removing BaseComponent. BaseActivity no longer requires one to be returned. Renaming inflation method to be more lifecycle centric. Creating a single scope which can be re-used for MVP classes. Removing HelloWorld module in favour of annotations. Re-instating removed disposable method for BasePresenter.
