+++ app/src/main/java/com/ustwo/boilerplate/application/ApplicationComponent.java
@@ -1,12 +1,11 @@
 package com.ustwo.boilerplate.application;
 
 import android.content.Context;
-import com.ustwo.boilerplate.base.BaseComponent;
 import dagger.Component;
 import javax.inject.Singleton;
 
 @Singleton
 @Component(modules = {ApplicationModule.class})
-public interface ApplicationComponent extends BaseComponent {
+public interface ApplicationComponent {
     Context getContext();
 }
