--- app/src/main/java/com/ustwo/boilerplate/base/BaseComponent.java
@@ -1,4 +1,0 @@
-package com.ustwo.boilerplate.base;
-
-public interface BaseComponent {
-}
