+++ app/src/main/java/com/ustwo/boilerplate/base/PerViewScope.java
@@ -1,4 +1,4 @@
-package com.ustwo.boilerplate.feature.helloworld;
+package com.ustwo.boilerplate.base;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -6,5 +6,5 @@ import javax.inject.Scope;
 
 @Scope
 @Retention(RetentionPolicy.RUNTIME)
-public @interface HelloWorldScope {
+public @interface PerViewScope {
 }
