+++ app/src/main/java/com/ustwo/boilerplate/feature/helloworld/HelloWorldComponent.java
@@ -1,11 +1,11 @@
 package com.ustwo.boilerplate.feature.helloworld;
 
 import com.ustwo.boilerplate.application.ApplicationComponent;
-import com.ustwo.boilerplate.base.BaseComponent;
+import com.ustwo.boilerplate.base.PerViewScope;
 import dagger.Component;
 
-@HelloWorldScope
-@Component(dependencies = ApplicationComponent.class, modules = HelloWorldModule.class)
-interface HelloWorldComponent extends BaseComponent {
+@PerViewScope
+@Component(dependencies = ApplicationComponent.class)
+interface HelloWorldComponent {
     void inject(HelloWorldActivity activity);
 }
