--- app/src/main/java/com/ustwo/boilerplate/feature/helloworld/HelloWorldModule.java
@@ -1,14 +1,0 @@
-package com.ustwo.boilerplate.feature.helloworld;
-
-import dagger.Module;
-import dagger.Provides;
-
-@Module
-class HelloWorldModule {
-
-    @HelloWorldScope
-    @Provides
-    HelloWorldPresenter provideHelloWorldPresenter() {
-        return new HelloWorldPresenter();
-    }
-}
